% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\name{activate_microbiome_dataset}
\alias{activate_microbiome_dataset}
\title{Activate a Specific Component of a Microbiome Dataset}
\usage{
activate_microbiome_dataset(
  .data,
  what = c("sample_info", "variable_info", "expression_data")
)
}
\arguments{
\item{.data}{The microbiome dataset object.}

\item{what}{A character string specifying the component to activate.
Options are "sample_info", "variable_info", or "expression_data".
Default is "sample_info".}
}
\value{
The modified microbiome dataset object with the specified component activated.
}
\description{
This function sets a specific component of a microbiome dataset object as active.
It allows the user to specify which part of the dataset to work with.
}
\details{
The function uses the \code{slot} function to assign the active component to
the 'activated' slot of the microbiome dataset object. This makes it easier to
work with specific parts of the dataset in subsequent analyses.
}
\note{
It's essential to ensure that the microbiome dataset object has the
necessary structure and slots to use this function effectively.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- activate_microbiome_dataset(global_patterns, what = "sample_info")
x@activated
}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
