% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{agglomerate_taxa}
\alias{agglomerate_taxa}
\title{Agglomerate Taxa to a Taxonomic Rank}
\usage{
agglomerate_taxa(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na_remove = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{A taxonomy rank in \code{variable_info}.}

\item{what}{Aggregation method.}

\item{na_remove}{Should taxa with missing rank be removed before agglomeration?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Collapse taxa in a \code{microbiome_dataset} to a taxonomy rank.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
x <- agglomerate_taxa(
  object = x,
  taxonomic_rank = "Genus",
  what = "sum_intensity"
)
dim(x@expression_data)
}
