% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{agglomerate_tree}
\alias{agglomerate_tree}
\title{Agglomerate Taxa Using Tree Distances}
\usage{
agglomerate_tree(
  object,
  k,
  h,
  what = c("sum_intensity", "mean_intensity", "median_intensity")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{k}{Number of clusters used to cut the tree.}

\item{h}{Height used to cut the tree.}

\item{what}{Aggregation method.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Collapse OTU-level abundance using the topology of \code{otu_tree}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
otu_phylo <- ape::rtree(
  n = nrow(x@variable_info),
  tip.label = x@variable_info$variable_id
)
x@otu_tree <- tidytree::treedata(
  phylo = otu_phylo,
  data = tibble::tibble(node = seq_len(length(otu_phylo$tip.label) + otu_phylo$Nnode))
)
x <- agglomerate_tree(x, k = 6)
dim(x@expression_data)
}
