% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{align_ref_seq}
\alias{align_ref_seq}
\title{Align Reference Sequences with the Current Dataset}
\usage{
align_ref_seq(object, drop_unmatched = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{drop_unmatched}{Should unmatched sequences be dropped? Defaults to
\code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Synchronize \code{ref_seq} with the current \code{variable_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
x@ref_seq <- Biostrings::DNAStringSet(rep("ACGT", nrow(x@variable_info)))
x <- align_ref_seq(x)
names(x@ref_seq)[1]
}
