% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{align_tree}
\alias{align_tree}
\title{Align a Tree Slot with the Current Dataset}
\usage{
align_tree(object, tree = c("otu_tree", "taxa_tree"), drop_unmatched = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot to align, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{drop_unmatched}{Should unmatched OTU tips be dropped when aligning the
OTU tree?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Synchronize \code{otu_tree} or \code{taxa_tree} with the current \code{variable_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
x <- align_tree(x, tree = "taxa_tree")
class(x@taxa_tree)
}
