% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ariththmetic.R
\name{log.microbiome_dataset}
\alias{log.microbiome_dataset}
\alias{abs.microbiome_dataset}
\alias{sqrt.microbiome_dataset}
\alias{scale.microbiome_dataset}
\alias{+,microbiome_dataset,numeric-method}
\alias{-,microbiome_dataset,numeric-method}
\alias{*,microbiome_dataset,numeric-method}
\alias{/,microbiome_dataset,numeric-method}
\alias{colSums,microbiome_dataset-method}
\alias{rowSums,microbiome_dataset-method}
\alias{colMeans,microbiome_dataset-method}
\alias{rowMeans,microbiome_dataset-method}
\title{log}
\usage{
\method{log}{microbiome_dataset}(x, base = exp(1))

\method{abs}{microbiome_dataset}(x)

\method{sqrt}{microbiome_dataset}(x)

\method{scale}{microbiome_dataset}(x, center = TRUE, scale = TRUE)

\S4method{+}{microbiome_dataset,numeric}(e1, e2)

\S4method{-}{microbiome_dataset,numeric}(e1, e2)

\S4method{*}{microbiome_dataset,numeric}(e1, e2)

\S4method{/}{microbiome_dataset,numeric}(e1, e2)

\S4method{colSums}{microbiome_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{rowSums}{microbiome_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{colMeans}{microbiome_dataset}(x, na.rm = FALSE, dims = 1)

\S4method{rowMeans}{microbiome_dataset}(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{x}

\item{base}{base}

\item{center}{center}

\item{scale}{scale}

\item{e1}{a microbiome_dataset class object}

\item{e2}{numeric}

\item{na.rm}{na.rm}

\item{dims}{dims}
}
\value{
microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

microbiome_dataset object

result

vector object

vector object

vector object
}
\description{
#' @title colSums
#' @method colSums microbiome_dataset
#' @param x x
#' @param na.rm na.rm
#' @param dims dims
#' @export
#' @rdname arithmetic-microbiome_dataset
#' @return vector object

#' @title rowSums
#' @method rowSums microbiome_dataset
#' @param x x
#' @param na.rm na.rm
#' @param dims dims
#' @export
#' @rdname arithmetic-microbiome_dataset
#' @return vector object

#' @title colMeans
#' @method colMeans microbiome_dataset
#' @param x x
#' @param na.rm na.rm
#' @param dims dims
#' @export
#' @rdname arithmetic-microbiome_dataset
#' @return vector object

#' @title rowMeans
#' @method rowMeans microbiome_dataset
#' @param x x
#' @param na.rm na.rm
#' @param dims dims
#' @export
#' @rdname arithmetic-microbiome_dataset
#' @return vector object
}
\details{
colSums.microbiome_dataset = function(x, na.rm = FALSE, dims = 1){
colSums(x@expression_data, na.rm = na.rm, dims = dims)
}

rowSums.microbiome_dataset = function(x, na.rm = FALSE, dims = 1){
rowSums(x@expression_data, na.rm = na.rm, dims = dims)
}

colMeans.microbiome_dataset = function(x, na.rm = FALSE, dims = 1){
colMeans(x@expression_data, na.rm = na.rm, dims = dims)
}

rowMeans.microbiome_dataset = function(x, na.rm = FALSE, dims = 1){
rowMeans(x@expression_data, na.rm = na.rm, dims = dims)
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
