% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{as_tibble_association}
\alias{as_tibble_association}
\title{Convert Association Results to a Tibble}
\usage{
as_tibble_association(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}
}
\value{
A tibble.
}
\description{
Convert Association Results to a Tibble
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
class(as_tibble_association(x))
}
