% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{as_tibble_differential_abundance}
\alias{as_tibble_differential_abundance}
\title{Convert Differential Abundance Results to a Tibble}
\usage{
as_tibble_differential_abundance(object)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}
}
\value{
A tibble.
}
\description{
Convert Differential Abundance Results to a Tibble
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
genus_object <- summarise_taxa(x0, taxonomic_rank = "Genus")
result_table <- data.frame(
  variable_id = genus_object@variable_info$variable_id[1:6],
  estimate = seq(-1, 1, length.out = 6)
)
x <- create_differential_abundance_result(
  result = cbind(result_table, genus_object@variable_info[1:6, ]),
  taxonomic_rank = "Genus"
)
class(as_tibble_differential_abundance(x))
}
