% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{as_tibble_diversity}
\alias{as_tibble_diversity}
\title{Convert Diversity Results to a Tibble}
\usage{
as_tibble_diversity(object)
}
\arguments{
\item{object}{A \code{microbiome_diversity} object.}
}
\value{
A tibble.
}
\description{
Convert a \code{microbiome_diversity} result to a tibble for tidy analysis.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- calculate_alpha_diversity(global_patterns, metric = "shannon")
y <- as_tibble_diversity(x)
class(y)
}
