% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{as_tibble_mechanism}
\alias{as_tibble_mechanism}
\title{Convert Mechanism Results to a Tibble}
\usage{
as_tibble_mechanism(object)
}
\arguments{
\item{object}{A \code{crossomics_mechanism} object.}
}
\value{
A tibble.
}
\description{
Convert Mechanism Results to a Tibble
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

pathway_link <- standardize_pathway_link(
  data.frame(
    taxon_id = summarise_taxa(
      demo_crossomics$microbiome_data,
      taxonomic_rank = "Genus"
    )@variable_info$variable_id[1],
    metabolite_id = demo_crossomics$metabolome_data@annotation_table$variable_id[1],
    pathway_id = "pathway_a",
    stringsAsFactors = FALSE
  )
)
x <- infer_metabolic_link(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  pathway_link = pathway_link,
  microbiome_rank = "Genus",
  q_value_cutoff = 1
)
class(as_tibble_mechanism(x))
}
