% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{as_tibble_network}
\alias{as_tibble_network}
\title{Convert Network Results to a Tibble}
\usage{
as_tibble_network(object, what = c("both", "nodes", "edges"))
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}

\item{what}{Which part of the result to convert.}
}
\value{
A tibble or named list of tibbles.
}
\description{
Convert Network Results to a Tibble
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
x <- build_correlation_network(x, min_abs_correlation = 0.2, max_q_value = 1)
class(as_tibble_network(x, what = "nodes"))
}
