% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{as_tibble_ordination}
\alias{as_tibble_ordination}
\title{Convert Ordination Results to a Tibble}
\usage{
as_tibble_ordination(object)
}
\arguments{
\item{object}{A \code{microbiome_ordination} object.}
}
\value{
A tibble.
}
\description{
Convert a \code{microbiome_ordination} result to a tibble for tidy analysis.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- run_ordination(global_patterns, method = "PCoA")
y <- as_tibble_ordination(x)
class(y)
}
