% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{associate_microbe_metabolite}
\alias{associate_microbe_metabolite}
\title{Associate Microbiome Taxa with Metabolites}
\usage{
associate_microbe_metabolite(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  microbiome_rank = "Genus",
  microbiome_transform = c("relative", "log", "presence_absence", "clr", "none"),
  metabolome_transform = c("log", "none"),
  method = c("spearman", "pearson", "sparcc"),
  p_adjust_method = "BH"
)
}
\arguments{
\item{object}{A \code{microbiome_metabolome_dataset} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object. Used when \code{object} is
not supplied.}

\item{metabolome_data}{A \code{mass_dataset} object. Used when \code{object} is not
supplied.}

\item{sample_link}{Optional sample link table used when \code{object} is not
supplied.}

\item{microbiome_rank}{Taxonomy rank for microbiome aggregation.}

\item{microbiome_transform}{Microbiome transformation method.}

\item{metabolome_transform}{Metabolome transformation method.}

\item{method}{Correlation method.}

\item{p_adjust_method}{Multiple-testing correction method.}
}
\value{
A \code{microbe_metabolite_association} object.
}
\description{
Compute pairwise microbe-metabolite associations on matched samples.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- create_microbiome_metabolome_dataset(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link
)
y <- associate_microbe_metabolite(x, microbiome_rank = "Genus")
class(y)
}
