% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{build_correlation_network}
\alias{build_correlation_network}
\title{Build a Microbe-Metabolite Correlation Network}
\usage{
build_correlation_network(
  object,
  min_abs_correlation = 0,
  max_q_value = 1,
  direction = c("any", "positive", "negative"),
  top_n = Inf,
  keep_isolated = FALSE
)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}

\item{min_abs_correlation}{Minimum absolute correlation estimate.}

\item{max_q_value}{Maximum q value.}

\item{direction}{Direction of correlation to keep.}

\item{top_n}{Maximum number of edges to keep.}

\item{keep_isolated}{Whether isolated nodes should be kept.}
}
\value{
A \code{microbe_metabolite_network} object.
}
\description{
Convert correlation results into a bipartite network object.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
y <- build_correlation_network(x, min_abs_correlation = 0.2, max_q_value = 0.5)
y
}
