% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{calculate_alpha_diversity}
\alias{calculate_alpha_diversity}
\title{Calculate Alpha Diversity}
\usage{
calculate_alpha_diversity(
  object,
  metric = c("observed", "shannon", "simpson", "pielou")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{metric}{Alpha diversity metric.}
}
\value{
A \code{microbiome_diversity} object.
}
\description{
Calculate a sample-wise alpha diversity metric from a \code{microbiome_dataset}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- calculate_alpha_diversity(global_patterns, metric = "shannon")
class(x)
head(x@result)
}
