% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{calculate_beta_diversity}
\alias{calculate_beta_diversity}
\title{Calculate Beta Diversity}
\usage{
calculate_beta_diversity(object, method = c("bray", "jaccard", "euclidean"))
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{method}{Distance method.}
}
\value{
A \code{microbiome_diversity} object.
}
\description{
Calculate a sample-wise beta diversity distance matrix from a
\code{microbiome_dataset}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- calculate_beta_diversity(global_patterns, method = "bray")
class(x)
dim(as.matrix(x@result))
}
