% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{calculate_correlation}
\alias{calculate_correlation}
\alias{calculate_correlation_spearman}
\alias{calculate_correlation_pearson}
\alias{calculate_correlation_sparcc}
\title{Calculate Microbe-Metabolite Correlations}
\usage{
calculate_correlation(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  microbiome_rank = "Genus",
  microbiome_transform = c("relative", "log", "presence_absence", "clr", "none"),
  metabolome_transform = c("log", "none"),
  method = c("spearman", "pearson", "sparcc"),
  p_adjust_method = "BH",
  min_taxa_prevalence = 0,
  min_taxa_abundance = 0,
  min_metabolite_prevalence = 0,
  min_metabolite_sd = 0
)

calculate_correlation_spearman(...)

calculate_correlation_pearson(...)

calculate_correlation_sparcc(...)
}
\arguments{
\item{object}{A \code{microbiome_metabolome_dataset} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object. Used when \code{object} is
not supplied.}

\item{metabolome_data}{A \code{mass_dataset} object. Used when \code{object} is not
supplied.}

\item{sample_link}{Optional sample link table used when \code{object} is not
supplied.}

\item{microbiome_rank}{Taxonomy rank for microbiome aggregation.}

\item{microbiome_transform}{Microbiome transformation method.}

\item{metabolome_transform}{Metabolome transformation method.}

\item{method}{Correlation method.}

\item{p_adjust_method}{Multiple-testing correction method.}

\item{min_taxa_prevalence}{Minimum number of non-zero samples required for a
microbiome feature.}

\item{min_taxa_abundance}{Minimum mean abundance required for a microbiome
feature.}

\item{min_metabolite_prevalence}{Minimum number of non-zero samples required
for a metabolite.}

\item{min_metabolite_sd}{Minimum standard deviation required for a
metabolite.}

\item{...}{Additional arguments passed through the convenience wrappers.}
}
\value{
A \code{microbe_metabolite_association} object.
}
\description{
Compute pairwise correlations between microbiome taxa and metabolites on
matched samples.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  microbiome_rank = "Genus",
  metabolome_transform = "none",
  method = "spearman"
)
class(x)
}
