% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{calculate_network_centrality}
\alias{calculate_network_centrality}
\title{Calculate Network Centrality}
\usage{
calculate_network_centrality(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}
}
\value{
A data.frame of node centrality metrics.
}
\description{
Add node centrality metrics to a correlation network.
}
