% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_microbiome_dataset.R
\name{check_microbiome_dataset}
\alias{check_microbiome_dataset}
\title{Validate the Format and Consistency of a Microbiome Dataset}
\usage{
check_microbiome_dataset(
  expression_data,
  sample_info,
  variable_info,
  otu_tree,
  taxa_tree,
  otu_tree_link,
  taxa_tree_link,
  ref_seq,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{A data frame representing expression data.}

\item{sample_info}{A data frame containing sample information.}

\item{variable_info}{A data frame containing variable information.}

\item{otu_tree}{An optional parameter for OTU tree data.}

\item{taxa_tree}{An optional parameter for taxa tree data.}

\item{otu_tree_link}{An optional explicit mapping between \code{variable_id} and
\code{otu_tree} labels.}

\item{taxa_tree_link}{An optional explicit mapping between \code{variable_id} and
\code{taxa_tree} labels.}

\item{ref_seq}{An optional parameter for reference sequence data.}

\item{sample_info_note}{A data frame containing notes on sample information.}

\item{variable_info_note}{A data frame containing notes on variable information.}
}
\value{
A string "all good." if all checks pass, otherwise returns a descriptive
error message indicating the issue found.
}
\description{
This function checks the structure and consistency of various components of a
microbiome dataset, including expression data, sample information, variable information,
and their respective notes.
}
\details{
The function performs a series of checks to ensure:
\itemize{
\item Required data frames (expression_data, sample_info, variable_info) are provided.
\item All data frames are of the correct class.
\item Key columns exist and there are no duplicated items in critical fields.
\item The dimensions of the data frames are consistent with each other.
Errors in these checks will result in descriptive error messages.
}
}
\examples{
data("global_patterns", package = "microbiomedataset")

check_microbiome_dataset(
  expression_data = global_patterns@expression_data,
  sample_info = global_patterns@sample_info,
  variable_info = global_patterns@variable_info,
  otu_tree = global_patterns@otu_tree,
  taxa_tree = global_patterns@taxa_tree,
  ref_seq = global_patterns@ref_seq,
  sample_info_note = global_patterns@sample_info_note,
  variable_info_note = global_patterns@variable_info_note
)
}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
