% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_microbiome_dataset.R
\name{check_microbiome_dataset_class}
\alias{check_microbiome_dataset_class}
\title{Check Integrity and Structure of Microbiome Dataset Object}
\usage{
check_microbiome_dataset_class(object)
}
\arguments{
\item{object}{A microbiome dataset object to be checked.}
}
\value{
TRUE if the dataset passes all checks, otherwise returns a character
vector listing the found errors.
}
\description{
Performs a series of checks on a microbiome dataset object to ensure its
integrity and correct structure. This includes verifying the format of
various components such as sample_info, variable_info, and expression_data.
}
\details{
The function performs checks to ensure:
\itemize{
\item \code{variable_info} has a \code{variable_id} column and no duplicated variable IDs.
\item \code{sample_info} has \code{sample_id} and \code{class} columns with no duplicated sample IDs.
\item \code{sample_info_note} and \code{variable_info_note} have appropriate columns if they exist.
\item Consistency between the \code{expression_data} and the \code{sample_info} and \code{variable_info}.
\item Consistency between \code{sample_info}, \code{variable_info} and their respective notes.
Any issues found are returned as character strings in a vector.
}
}
\examples{
data("global_patterns", package = "microbiomedataset")

check_microbiome_dataset_class(global_patterns)
}
\author{
Xiaotao Shen \email{xiaotao.shen@outlook.com}
}
