% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_microbiome_dataset.R
\name{convert2microbiome_dataset}
\alias{convert2microbiome_dataset}
\alias{as.microbiome_dataset}
\alias{convert2microbiome_dataset.phyloseq}
\alias{convert2microbiome_dataset.mass_dataset}
\alias{convert2microbiome_dataset.TreeSummarizedExperiment}
\alias{convert2microbiome_dataset.MPSE}
\title{convert to microbiome_dataset class object.}
\usage{
convert2microbiome_dataset(object, ...)

as.microbiome_dataset(object, ...)

\method{convert2microbiome_dataset}{phyloseq}(object, ...)

\method{convert2microbiome_dataset}{mass_dataset}(object, ...)

\method{convert2microbiome_dataset}{TreeSummarizedExperiment}(object, assay_name = NULL, ...)

\method{convert2microbiome_dataset}{MPSE}(object, ...)
}
\arguments{
\item{object}{phylosep or other class}

\item{...}{other params}

\item{assay_name}{Assay name used when \code{object} is a
\code{TreeSummarizedExperiment}.}
}
\value{
microbiome_dataset class object.
}
\description{
convert to microbiome_dataset class object.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
phyloseq_object <- convert2phyloseq(x0)
x <- convert2microbiome_dataset(phyloseq_object)
class(x)
}
