% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_microbiome_dataset.R
\name{convert2MPSE}
\alias{convert2MPSE}
\alias{as.MPSE}
\alias{convert2MPSE.microbiome_dataset}
\title{convert to MPSE class object.}
\usage{
convert2MPSE(object, ...)

as.MPSE(object, ...)

\method{convert2MPSE}{microbiome_dataset}(object, ...)
}
\arguments{
\item{object}{microbiomedataset}

\item{...}{other params}
}
\value{
MPSE class object.
}
\description{
convert to MPSE class object.
}
\examples{
if (requireNamespace("MicrobiotaProcess", quietly = TRUE)) {
  data("global_patterns", package = "microbiomedataset")
  x <- convert2MPSE(global_patterns)
  class(x)
}
}
