% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_microbiome_dataset.R
\name{convert2phyloseq}
\alias{convert2phyloseq}
\alias{as.phyloseq}
\alias{convert2phyloseq.microbiome_dataset}
\title{convert to phyloseq class object.}
\usage{
convert2phyloseq(object, ...)

as.phyloseq(object, ...)

\method{convert2phyloseq}{microbiome_dataset}(object, ...)
}
\arguments{
\item{object}{microbiomedataset}

\item{...}{other params}
}
\value{
phyloseq class object.
}
\description{
convert to phyloseq class object.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- convert2phyloseq(global_patterns)
class(x)
}
