% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_microbiome_dataset.R
\name{convert2tbl_graph}
\alias{convert2tbl_graph}
\alias{as.tbl_graph}
\alias{convert2tbl_graph.microbiome_dataset}
\title{convert to tbl_graph class object in tidygraph.}
\usage{
convert2tbl_graph(object, intensity_method = c("sum", "mean", "median"), ...)

as.tbl_graph(object, intensity_method = c("sum", "mean", "median"), ...)

\method{convert2tbl_graph}{microbiome_dataset}(object, intensity_method = c("sum", "mean", "median"), ...)
}
\arguments{
\item{object}{microbiomedataset or Taxa table}

\item{intensity_method}{for each node, the how to calculate the intensity for
it accross all the samples}

\item{...}{other params}
}
\value{
tbl_graph class object.
}
\description{
convert to tbl_graph class object in tidygraph.
}
