% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_microbiome_dataset.R
\name{convert2TreeSummarizedExperiment}
\alias{convert2TreeSummarizedExperiment}
\alias{as.TreeSummarizedExperiment}
\alias{convert2TreeSummarizedExperiment.microbiome_dataset}
\title{convert to TreeSummarizedExperiment class object.}
\usage{
convert2TreeSummarizedExperiment(object, ...)

as.TreeSummarizedExperiment(object, ...)

\method{convert2TreeSummarizedExperiment}{microbiome_dataset}(object, assay_name = "counts", ...)
}
\arguments{
\item{object}{microbiomedataset}

\item{...}{other params}

\item{assay_name}{assay name used in the output object.}
}
\value{
TreeSummarizedExperiment class object.
}
\description{
convert to TreeSummarizedExperiment class object.
}
\examples{
if (requireNamespace("TreeSummarizedExperiment", quietly = TRUE)) {
  data("global_patterns", package = "microbiomedataset")
  x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
  x <- convert2TreeSummarizedExperiment(x0)
  class(x)
}
}
