% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{convert_network_graph}
\alias{convert_network_graph}
\title{Convert a Correlation Network to tbl_graph}
\usage{
convert_network_graph(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}
}
\value{
A \code{tbl_graph} object.
}
\description{
Convert a Correlation Network to tbl_graph
}
