% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{methods}
\name{create_microbiome_dataset}
\alias{create_microbiome_dataset}
\title{create_microbiome_dataset}
\usage{
create_microbiome_dataset(
  expression_data,
  sample_info,
  variable_info,
  otu_tree,
  taxa_tree,
  otu_tree_link,
  taxa_tree_link,
  ref_seq,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{MS1 peak table name.}

\item{sample_info}{Sample information name.}

\item{variable_info}{MS1 peak table name.
Columns are samples and rows are variables.}

\item{otu_tree}{otu_tree}

\item{taxa_tree}{taxa_tree}

\item{otu_tree_link}{explicit mapping between \code{variable_id} and \code{otu_tree}.}

\item{taxa_tree_link}{explicit mapping between \code{variable_id} and \code{taxa_tree}.}

\item{ref_seq}{ref_seq}

\item{sample_info_note}{Sample information name.}

\item{variable_info_note}{Sample information name.}
}
\value{
A microbiome_dataset-class object.
}
\description{
Create the microbiome_dataset object.
}
\examples{

expression_data <-
  as.data.frame(matrix(
    sample(1:100, 100, replace = TRUE),
    nrow = 10,
    ncol = 10
  ))

rownames(expression_data) <- paste0("OTU", 1:nrow(expression_data))
colnames(expression_data) <-
  paste0("Sample", 1:ncol(expression_data))
expression_data

variable_info <-
  as.data.frame(matrix(
    sample(letters, 70, replace = TRUE),
    nrow = nrow(expression_data),
    ncol = 7
  ))

rownames(variable_info) <- rownames(expression_data)
colnames(variable_info) <-
  c("Domain",
    "Phylum",
    "Class",
    "Order",
    "Family",
    "Genus",
    "Species")

variable_info$variable_id <-
  rownames(expression_data)

sample_info <-
  data.frame(sample_id = colnames(expression_data),
             class = "Subject")

object <-
  create_microbiome_dataset(
    expression_data = expression_data,
    sample_info = sample_info,
    variable_info = variable_info
  )

object
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
