% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{crossomics_integration-class}
\alias{crossomics_integration-class}
\title{crossomics_integration class}
\description{
crossomics_integration class
}
\section{Slots}{

\describe{
\item{\code{sample_coord}}{A data.frame of integration sample coordinates.}

\item{\code{microbiome_loading}}{A data.frame of microbiome feature loadings.}

\item{\code{metabolome_loading}}{A data.frame of metabolome feature loadings.}

\item{\code{sample_info}}{A data.frame of paired sample metadata.}

\item{\code{method}}{Integration method.}

\item{\code{microbiome_rank}}{Taxonomic level used for microbiome features.}

\item{\code{ncomp}}{Number of latent components.}
}}

