% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{crossomics_mechanism-class}
\alias{crossomics_mechanism-class}
\title{crossomics_mechanism class}
\description{
crossomics_mechanism class
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame of mechanism link results.}

\item{\code{method}}{Mechanism inference method.}

\item{\code{evidence_type}}{Evidence type used in the result.}
}}

