% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{detect_network_modules}
\alias{detect_network_modules}
\title{Detect Network Modules}
\usage{
detect_network_modules(object, method = c("louvain", "walktrap"))
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}

\item{method}{Community detection method.}
}
\value{
A data.frame of node-module assignments.
}
\description{
Run community detection on a microbe-metabolite network.
}
