% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{differential_abundance_result-class}
\alias{differential_abundance_result-class}
\title{differential_abundance_result class}
\description{
differential_abundance_result class
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame of differential abundance results.}

\item{\code{method}}{Differential abundance method.}

\item{\code{taxonomic_rank}}{Taxonomic level used in the result.}

\item{\code{group}}{Comparison or grouping label.}
}}

