% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{export_ref_seq}
\alias{export_ref_seq}
\title{Export Reference Sequences to FASTA}
\usage{
export_ref_seq(object, file, width = 80)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{file}{Output FASTA path.}

\item{width}{Number of letters per line in the FASTA file.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write \code{ref_seq} from a \code{microbiome_dataset} to a FASTA file for use in
external tools.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
ref_seq <- Biostrings::DNAStringSet(rep("ACGT", nrow(x@variable_info)))
x <- replace_ref_seq(x, value = ref_seq)
fasta_file <- tempfile(fileext = ".fasta")
export_ref_seq(x, fasta_file)
file.exists(fasta_file)
}
