% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\docType{methods}
\name{extract_otu_tree}
\alias{extract_otu_tree}
\alias{extract_taxa_tree}
\alias{extract_ref_seq}
\title{extract_otu_tree}
\usage{
extract_otu_tree(object)

extract_taxa_tree(object)

extract_ref_seq(object)
}
\arguments{
\item{object}{(required) microbiome_dataset class object.}
}
\value{
A treedata

A treedata

A xstringset
}
\description{
Extract OTU tree.

Extract Taxa tree.

Extract ref sequence.
}
\examples{
data("global_patterns", package = "microbiomedataset")

otu_tree <- extract_otu_tree(global_patterns)
class(otu_tree)
data("global_patterns", package = "microbiomedataset")

taxa_tree <- extract_taxa_tree(global_patterns)
class(taxa_tree)
data("global_patterns", package = "microbiomedataset")

ref_seq <- extract_ref_seq(global_patterns)
class(ref_seq)
}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
