% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{extract_association_result}
\alias{extract_association_result}
\title{Extract Cross-Omics Association Results}
\usage{
extract_association_result(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}
}
\value{
A data.frame.
}
\description{
Return the tidy association table from a
\code{microbe_metabolite_association} object.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
head(extract_association_result(x))
}
