% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{extract_correlation_table}
\alias{extract_correlation_table}
\title{Extract a Tidy Correlation Table}
\usage{
extract_correlation_table(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}
}
\value{
A data.frame.
}
\description{
Return the tidy result table from a correlation result object.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
head(extract_correlation_table(x))
}
