% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_extractors.R
\name{extract_differential_abundance_result}
\alias{extract_differential_abundance_result}
\title{Extract Differential Abundance Results}
\usage{
extract_differential_abundance_result(object)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}
}
\value{
A data.frame.
}
\description{
Return the tidy result table from a \code{differential_abundance_result} object.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
genus_object <- summarise_taxa(x0, taxonomic_rank = "Genus")
result_table <- data.frame(
  variable_id = genus_object@variable_info$variable_id[1:6],
  estimate = seq(-1, 1, length.out = 6)
)
x <- create_differential_abundance_result(
  result = cbind(result_table, genus_object@variable_info[1:6, ]),
  taxonomic_rank = "Genus"
)
head(extract_differential_abundance_result(x))
}
