% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{extract_diversity_result}
\alias{extract_diversity_result}
\title{Extract Diversity Results}
\usage{
extract_diversity_result(object)
}
\arguments{
\item{object}{A \code{microbiome_diversity} object.}
}
\value{
A data.frame or \code{dist} object.
}
\description{
Extract the tabular or distance result stored in a \code{microbiome_diversity}
object.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- calculate_alpha_diversity(global_patterns, metric = "shannon")
y <- extract_diversity_result(x)
class(y)
}
