% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{extract_network_table}
\alias{extract_network_table}
\title{Extract Network Tables}
\usage{
extract_network_table(object)
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}
}
\value{
A named list with \code{nodes} and \code{edges}.
}
\description{
Return node and edge tables from a network object.
}
