% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{extract_ordination_result}
\alias{extract_ordination_result}
\title{Extract Ordination Coordinates}
\usage{
extract_ordination_result(object)
}
\arguments{
\item{object}{A \code{microbiome_ordination} object.}
}
\value{
A data.frame.
}
\description{
Extract sample coordinates from a \code{microbiome_ordination} object.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- run_ordination(global_patterns, method = "PCoA")
y <- extract_ordination_result(x)
head(y)
}
