% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{extract_tree_data}
\alias{extract_tree_data}
\title{Extract Tree Data from a Microbiome Dataset}
\usage{
extract_tree_data(
  object,
  tree = c("otu_tree", "taxa_tree"),
  data_type = c("treedata", "phylo", "table")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot to extract, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{data_type}{Output type, one of \code{"treedata"}, \code{"phylo"}, or \code{"table"}.}
}
\value{
A tree object or data frame depending on \code{data_type}.
}
\description{
Export \code{otu_tree} or \code{taxa_tree} from a \code{microbiome_dataset} in the desired
representation.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
tree_tbl <- extract_tree_data(x, tree = "taxa_tree", data_type = "table")
head(tree_tbl)
}
