% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{extract_tree_link_data}
\alias{extract_tree_link_data}
\title{Extract Tree Link Metadata}
\usage{
extract_tree_link_data(object, tree = c("otu_tree", "taxa_tree"))
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot, either \code{"otu_tree"} or \code{"taxa_tree"}.}
}
\value{
A data frame or \code{NULL}.
}
\description{
Return the explicit mapping table between \code{variable_id} and nodes in
\code{otu_tree} or \code{taxa_tree}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
extract_tree_link_data(x, tree = "taxa_tree")[1:6, ]
}
