% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{filter_correlation}
\alias{filter_correlation}
\title{Filter Correlation Results}
\usage{
filter_correlation(
  object,
  min_abs_correlation = 0,
  max_q_value = 1,
  direction = c("any", "positive", "negative")
)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}

\item{min_abs_correlation}{Minimum absolute correlation estimate.}

\item{max_q_value}{Maximum q value.}

\item{direction}{Direction of correlation to keep.}
}
\value{
A \code{microbe_metabolite_association} object.
}
\description{
Keep only correlation edges that pass user-defined score and significance
thresholds.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  metabolome_transform = "none"
)
y <- filter_correlation(x, min_abs_correlation = 0.3, max_q_value = 0.2)
y
}
