% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter Samples by Library Size}
\usage{
filter_samples(object, min_abundance = NULL, max_abundance = NULL)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{min_abundance}{Minimum total abundance required to keep a sample.}

\item{max_abundance}{Maximum total abundance allowed to keep a sample.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Filter Samples by Library Size
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- filter_samples(global_patterns, min_abundance = 1000)
dim(x@expression_data)
}
