% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{filter_samples_by_group}
\alias{filter_samples_by_group}
\title{Filter Samples by a Sample Metadata Group}
\usage{
filter_samples_by_group(object, group_by, groups, keep = TRUE, na_keep = FALSE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{group_by}{A column in \code{sample_info}.}

\item{groups}{Character vector of group labels used for filtering.}

\item{keep}{Should the provided groups be kept? Defaults to \code{TRUE}.}

\item{na_keep}{Should samples with missing group values be kept? Defaults to
\code{FALSE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Keep or remove samples using values from a column in \code{sample_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- filter_samples_by_group(
  global_patterns,
  group_by = "SampleType",
  groups = c("Soil", "Ocean")
)
dim(x@expression_data)
}
