% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter Taxa by Abundance or Prevalence}
\usage{
filter_taxa(
  object,
  min_prevalence = NULL,
  min_prevalence_prop = NULL,
  min_abundance = NULL,
  detection = 0
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{min_prevalence}{Minimum prevalence required to keep a taxon.}

\item{min_prevalence_prop}{Minimum prevalence proportion required to keep a taxon.}

\item{min_abundance}{Minimum total abundance required to keep a taxon.}

\item{detection}{Detection threshold used for prevalence.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Filter Taxa by Abundance or Prevalence
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- filter_taxa(
  global_patterns,
  min_prevalence = 5,
  min_abundance = 100
)
dim(x@expression_data)
}
