% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{filter_taxa_by_rank}
\alias{filter_taxa_by_rank}
\title{Filter Taxa by a Taxonomic Rank}
\usage{
filter_taxa_by_rank(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  taxa,
  keep = TRUE,
  na_keep = FALSE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{A taxonomy rank in \code{variable_info}.}

\item{taxa}{Character vector of taxa labels used for filtering.}

\item{keep}{Should the provided taxa be kept? Defaults to \code{TRUE}.}

\item{na_keep}{Should taxa with missing rank be kept? Defaults to \code{FALSE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Keep or remove taxa using values in a taxonomy column from \code{variable_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- filter_taxa_by_rank(
  global_patterns,
  taxonomic_rank = "Phylum",
  taxa = c("Bacteroidetes", "Firmicutes")
)
dim(x@expression_data)
}
