% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{filter_tree_by_rank}
\alias{filter_tree_by_rank}
\title{Filter the Taxa Tree by a Taxonomic Rank}
\usage{
filter_tree_by_rank(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  taxa,
  keep = TRUE,
  na_keep = FALSE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{A taxonomy rank in \code{variable_info}.}

\item{taxa}{Character vector of taxa labels used for filtering.}

\item{keep}{Should the provided taxa be kept? Defaults to \code{TRUE}.}

\item{na_keep}{Should taxa with missing rank be kept? Defaults to \code{FALSE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Rebuild \code{taxa_tree} using only taxa that match selected labels at a chosen
taxonomy rank, without changing the abundance matrix.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
x <- filter_tree_by_rank(
  x,
  taxonomic_rank = "Phylum",
  taxa = c("Bacteroidetes", "Firmicutes")
)
class(x@taxa_tree)
}
