% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_logo.R
\docType{methods}
\name{get_microbiomedataset_version}
\alias{get_microbiomedataset_version}
\title{Retrieve the Version of the microbiomedataset Package}
\usage{
get_microbiomedataset_version()
}
\value{
A character string representing the version of the
microbiomedataset package.
}
\description{
This function returns the current version of the microbiomedataset package.
}
\details{
\code{get_microbiomedataset_version} retrieves the version number of the
microbiomedataset package using the \code{packageVersion} function from the utils
package. It returns the version as a character string.
}
\note{
This function is especially useful for logging, debugging, or
providing user feedback regarding the version of the microbiomedataset package.
}
\examples{
get_microbiomedataset_version()

}
\author{
Xiaotao Shen
\email{xiaotao.shen@outlook.com}
}
