% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{import_ref_seq}
\alias{import_ref_seq}
\title{Import Reference Sequences from FASTA}
\usage{
import_ref_seq(object, file, format = "fasta", align = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{file}{Input FASTA path.}

\item{format}{Sequence format passed to \code{\link[Biostrings:XStringSet-io]{Biostrings::readDNAStringSet()}}.}

\item{align}{Should imported sequences be aligned to \code{variable_info}?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Read a FASTA file into \code{ref_seq} and optionally align sequence names to the
current dataset.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
fasta_file <- tempfile(fileext = ".fasta")
ref_seq <- Biostrings::DNAStringSet(rep("ACGT", nrow(x@variable_info)))
names(ref_seq) <- x@variable_info$variable_id
Biostrings::writeXStringSet(ref_seq, fasta_file, format = "fasta")
x <- import_ref_seq(x, fasta_file)
length(x@ref_seq)
}
