% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{import_tree}
\alias{import_tree}
\title{Import a Tree from Newick into a Microbiome Dataset}
\usage{
import_tree(
  object,
  file,
  tree = c("otu_tree", "taxa_tree"),
  format = c("newick"),
  align = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{file}{Input tree path.}

\item{tree}{Tree slot to import into, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{format}{Tree format. Currently only \code{"newick"} is supported.}

\item{align}{Should the imported tree be aligned to the current dataset?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Read a tree file and store it in \code{otu_tree} or \code{taxa_tree}, optionally
aligning the imported tree to the current dataset.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
tree_file <- tempfile(fileext = ".nwk")
otu_phylo <- ape::rtree(
  n = nrow(x@variable_info),
  tip.label = x@variable_info$variable_id
)
ape::write.tree(otu_phylo, file = tree_file)
x <- import_tree(x, file = tree_file, tree = "otu_tree")
class(x@otu_tree)
}
