% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{integrate_microbe_metabolite}
\alias{integrate_microbe_metabolite}
\title{Integrate Matched Microbiome and Metabolome Data}
\usage{
integrate_microbe_metabolite(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  method = c("spls", "diablo"),
  microbiome_rank = "Genus",
  ncomp = 2
)
}
\arguments{
\item{object}{A \code{microbiome_metabolome_dataset} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object. Used when \code{object} is
not supplied.}

\item{metabolome_data}{A \code{mass_dataset} object. Used when \code{object} is not
supplied.}

\item{sample_link}{Optional sample link table used when \code{object} is not
supplied.}

\item{method}{Integration method.}

\item{microbiome_rank}{Taxonomy rank for microbiome aggregation.}

\item{ncomp}{Number of latent components.}
}
\value{
A \code{crossomics_integration} object.
}
\description{
Wrapper for multiblock integration methods on matched omics matrices.
}
\examples{
data("global_patterns", package = "microbiomedataset")

data("demo_crossomics", package = "microbiomedataset")

if (requireNamespace("mixOmics", quietly = TRUE)) {
  x <- integrate_microbe_metabolite(
    microbiome_data = demo_crossomics$microbiome_data,
    metabolome_data = demo_crossomics$metabolome_data,
    sample_link = demo_crossomics$sample_link,
    method = "spls"
  )
  class(x)
}
}
