% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{melt_crossomics_features}
\alias{melt_crossomics_features}
\title{Melt Matched Cross-Omics Features}
\usage{
melt_crossomics_features(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  microbiome_rank = NULL
)
}
\arguments{
\item{object}{A \code{microbiome_metabolome_dataset} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object. Used when \code{object} is
not supplied.}

\item{metabolome_data}{A \code{mass_dataset} object. Used when \code{object} is not
supplied.}

\item{sample_link}{Optional sample link table used when \code{object} is not
supplied.}

\item{microbiome_rank}{Optional taxonomy rank used to aggregate microbiome
abundance before melting.}
}
\value{
A list with \code{microbiome} and \code{metabolome} tibbles.
}
\description{
Export matched microbiome and metabolome data as tidy long tables.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- create_microbiome_metabolome_dataset(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link
)
y <- melt_crossomics_features(x, microbiome_rank = "Genus")
names(y)
}
