% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abundance.R
\name{melt_features}
\alias{melt_features}
\title{Melt Features into a Long Table}
\usage{
melt_features(object, relative = FALSE, value_name = "abundance")
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{relative}{Should abundance be converted to relative abundance before
melting?}

\item{value_name}{Name of the abundance column in the output.}
}
\value{
A data frame in long format.
}
\description{
Neutral alias of \code{\link[=psmelt_microbiome_dataset]{psmelt_microbiome_dataset()}} for feature-level long table
export.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- melt_features(global_patterns)
head(x[, c("sample_id", "variable_id", "abundance")])
}
