% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{melt_tree}
\alias{melt_tree}
\title{Melt a Tree Slot into a Tidy Table}
\usage{
melt_tree(object, tree = c("otu_tree", "taxa_tree"), include_tree = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot to melt, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{include_tree}{Should the output include a \code{tree} column? Defaults to
\code{TRUE}.}
}
\value{
A data frame.
}
\description{
Convert \code{otu_tree} or \code{taxa_tree} into a tidy table for inspection,
visualisation, or downstream joins.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
tree_tbl <- melt_tree(x, tree = "taxa_tree")
head(tree_tbl)
}
