% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{merge_samples}
\alias{merge_samples}
\title{Merge Samples by a Sample Metadata Column}
\usage{
merge_samples(
  object,
  group_by,
  what = c("sum_intensity", "mean_intensity", "median_intensity")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{group_by}{A column in \code{sample_info}.}

\item{what}{Aggregation method.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Merge samples in a \code{microbiome_dataset} using a column in \code{sample_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- merge_samples(
  object = global_patterns,
  group_by = "SampleType",
  what = "sum_intensity"
)
dim(x@expression_data)
}
