% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbe_metabolite_association-class}
\alias{microbe_metabolite_association-class}
\title{microbe_metabolite_association class}
\description{
microbe_metabolite_association class
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame of association results.}

\item{\code{method}}{Association method.}

\item{\code{microbiome_rank}}{Taxonomic level used for microbiome features.}

\item{\code{metabolome_transform}}{Transformation applied to metabolome data.}
}}

