% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbe_metabolite_network-class}
\alias{microbe_metabolite_network-class}
\title{microbe_metabolite_network class}
\description{
microbe_metabolite_network class
}
\section{Slots}{

\describe{
\item{\code{node_data}}{A data.frame of network nodes.}

\item{\code{edge_data}}{A data.frame of network edges.}

\item{\code{method}}{Correlation method used to derive the network.}

\item{\code{microbiome_rank}}{Taxonomic level used for microbiome features.}

\item{\code{thresholds}}{A list describing edge filtering thresholds.}
}}

