% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbiome_dataset-class}
\alias{microbiome_dataset-class}
\title{microbiome_dataset class}
\description{
microbiome_dataset class
}
\section{Slots}{

\describe{
\item{\code{otu_tree}}{A treedata object of tidytree package or NULL.}

\item{\code{taxa_tree}}{A treedata object of tidytree package or NULL.}

\item{\code{otu_tree_link}}{A data.frame linking \code{variable_id} to \code{otu_tree} nodes.}

\item{\code{taxa_tree_link}}{A data.frame linking \code{variable_id} to \code{taxa_tree} nodes.}

\item{\code{ref_seq}}{A XStringSet object of Biostrings package or NULL.}

\item{\code{...}}{Other slots from mass_dataset class object in massdataset pacakge}
}}

