% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbiome_diversity-class}
\alias{microbiome_diversity-class}
\title{microbiome_diversity class}
\description{
microbiome_diversity class
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame or matrix containing diversity results.}

\item{\code{type}}{Diversity type, such as alpha or beta.}

\item{\code{method}}{Method used for diversity calculation.}
}}

