% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbiome_metabolome_dataset-class}
\alias{microbiome_metabolome_dataset-class}
\title{microbiome_metabolome_dataset class}
\description{
microbiome_metabolome_dataset class
}
\section{Slots}{

\describe{
\item{\code{microbiome_data}}{A \code{microbiome_dataset} object.}

\item{\code{metabolome_data}}{A \code{mass_dataset} object.}

\item{\code{sample_link}}{A data.frame linking microbiome and metabolome samples.}

\item{\code{feature_link}}{Optional taxon-metabolite link table.}

\item{\code{pathway_link}}{Optional pathway/reaction link table.}

\item{\code{process_info}}{Process tracking metadata.}

\item{\code{version}}{Package version used to build the object.}
}}

