% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{microbiome_ordination-class}
\alias{microbiome_ordination-class}
\title{microbiome_ordination class}
\description{
microbiome_ordination class
}
\section{Slots}{

\describe{
\item{\code{sample_coord}}{A data.frame of sample coordinates.}

\item{\code{eigenvalues}}{Numeric vector of eigenvalues or explained variance.}

\item{\code{method}}{Ordination method.}

\item{\code{distance_method}}{Distance method used, if applicable.}

\item{\code{feature_loading}}{A data.frame of feature loadings, if available.}
}}

