% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_conflicts.R
\docType{methods}
\name{microbiomedataset_conflicts}
\alias{microbiomedataset_conflicts}
\title{Conflicts between the microbiomedataset and other packages}
\usage{
microbiomedataset_conflicts()
}
\value{
microbiomedataset conflicts
}
\description{
This function lists all the conflicts between packages in the microbiomedataset
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
microbiomedataset_conflicts()
}
