% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_microbiome.R
\name{microbiomedataset_packages}
\alias{microbiomedataset_packages}
\title{List all packages in the microbiomedataset}
\usage{
microbiomedataset_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include microbiomedataset in the list?}
}
\value{
microbiomedataset packages
}
\description{
List all packages in the microbiomedataset
}
\examples{
microbiomedataset_packages()
}
