% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\docType{class}
\name{mixed_effect_result_class}
\alias{mixed_effect_result_class}
\alias{longitudinal_mixed_effect_result-class}
\alias{longitudinal_mixed_effect_result}
\title{longitudinal_mixed_effect_result class}
\description{
longitudinal_mixed_effect_result class
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame of longitudinal mixed-effect results.}

\item{\code{method}}{Modeling method label.}

\item{\code{time_variable}}{Time variable used in the model.}

\item{\code{group_variable}}{Optional grouping variable.}
}}

