% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sample.R
\name{mutate2sample}
\alias{mutate2sample}
\alias{mutate2sample.microbiome_dataset}
\title{Mutate new columns into sample_info}
\usage{
mutate2sample(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop"),
  according_to_variables = "all",
  ...
)

\method{mutate2sample}{microbiome_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity", "na_number", "na_prop"),
  according_to_variables = "all",
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{what}{which you want to mutate}

\item{according_to_variables}{(required) What variables used to calculate.
Default is "all". If you
want to use only several variables, provide their names as a vector.}

\item{...}{other params}
}
\value{
microbiome_dataset or data.frame.
}
\description{
Mutate new columns into sample_info
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- mutate2sample(global_patterns, what = "sum_intensity")
head(x@sample_info)
}
