% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot Taxon Abundance}
\usage{
plot_abundance(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  taxa = NULL,
  top_n = 10,
  x = "sample_id",
  geom = c("boxplot", "violin", "point"),
  relative = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{taxa}{Optional taxa to keep. Defaults to the most abundant taxa.}

\item{top_n}{Number of taxa shown when \code{taxa} is not supplied.}

\item{x}{Sample metadata column shown on the x axis.}

\item{geom}{Geometry used to display abundance.}

\item{relative}{Whether to show relative abundance.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw sample-wise abundance distributions for selected taxa.
}
\examples{
data("global_patterns", package = "microbiomedataset")

plot_abundance(
  object = global_patterns,
  taxonomic_rank = "Phylum",
  x = "SampleType",
  top_n = 6
)
}
