% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_abundance_density}
\alias{plot_abundance_density}
\title{Plot Abundance Density}
\usage{
plot_abundance_density(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  relative = TRUE,
  log10_transform = FALSE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{top_n}{Number of taxa displayed.}

\item{relative}{Whether to use relative abundance.}

\item{log10_transform}{Should abundance values be log10 transformed with a pseudocount?}
}
\value{
A \code{ggplot} object.
}
\description{
Plot abundance distributions for dominant taxa.
}
\examples{
data("global_patterns", package = "microbiomedataset")
plot_abundance_density(global_patterns, taxonomic_rank = "Phylum", top_n = 5)
}
