% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{Plot Beta Diversity}
\usage{
plot_beta_diversity(object, annotate_by = NULL, cluster = FALSE)
}
\arguments{
\item{object}{A \code{microbiome_diversity} object with \code{type = "beta"}.}

\item{annotate_by}{Optional named character vector used to annotate samples.
Names must be sample identifiers.}

\item{cluster}{Should samples be reordered by hierarchical clustering?}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a heatmap-style plot from beta diversity distance results.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- calculate_beta_diversity(global_patterns, method = "bray")
p <- plot_beta_diversity(x)
class(p)
}
