% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization.R
\name{plot_composition}
\alias{plot_composition}
\title{Plot Microbiome Composition}
\usage{
plot_composition(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  x = "sample_id",
  relative = TRUE,
  facet_by = NULL
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank displayed in the stack.}

\item{top_n}{Number of dominant taxa shown per sample.}

\item{x}{Sample metadata column shown on the x axis.}

\item{relative}{Whether to show relative abundance.}

\item{facet_by}{Optional sample metadata column used for faceting.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a stacked composition barplot at a selected taxonomic rank.
}
\examples{
data("global_patterns", package = "microbiomedataset")

plot_composition(
  object = global_patterns,
  taxonomic_rank = "Phylum",
  top_n = 8,
  x = "SampleType"
)
}
