% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_core_taxa}
\alias{plot_core_taxa}
\title{Plot Core Taxa}
\usage{
plot_core_taxa(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  detection = 0,
  prevalence_cutoff = 0.5,
  abundance_cutoff = NULL,
  relative = TRUE,
  top_n = 20
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{detection}{Detection threshold used for prevalence.}

\item{prevalence_cutoff}{Core prevalence cutoff.}

\item{abundance_cutoff}{Core mean abundance cutoff.}

\item{relative}{Whether to use relative abundance.}

\item{top_n}{Number of taxa displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Highlight taxa that satisfy prevalence and abundance thresholds.
}
\examples{
data("global_patterns", package = "microbiomedataset")
x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
plot_core_taxa(x, taxonomic_rank = "Phylum", top_n = 8)
}
