% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization_extended.R
\name{plot_correlation_dotplot}
\alias{plot_correlation_dotplot}
\title{Plot Correlations as a Dotplot}
\usage{
plot_correlation_dotplot(object, top_n_taxa = 20, top_n_metabolites = 20)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}

\item{top_n_taxa}{Number of taxa displayed.}

\item{top_n_metabolites}{Number of metabolites displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a bubble plot of top microbe-metabolite correlations.
}
