% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{Plot a Correlation Heatmap}
\usage{
plot_correlation_heatmap(
  object,
  top_n_taxa = 20,
  top_n_metabolites = 20,
  value = c("estimate", "q_value")
)
}
\arguments{
\item{object}{A \code{microbe_metabolite_association} object.}

\item{top_n_taxa}{Number of taxa retained.}

\item{top_n_metabolites}{Number of metabolites retained.}

\item{value}{Column mapped to fill.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a heatmap of top microbe-metabolite correlations.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

correlation <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  microbiome_rank = "Genus",
  metabolome_transform = "none"
)
plot_correlation_heatmap(correlation, top_n_taxa = 8, top_n_metabolites = 8)
}
