% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{plot_crossomics_loading_heatmap}
\alias{plot_crossomics_loading_heatmap}
\title{Plot Cross-Omics Loading Heatmap}
\usage{
plot_crossomics_loading_heatmap(
  object,
  block = c("microbiome", "metabolome"),
  axes = NULL,
  top_n = 20
)
}
\arguments{
\item{object}{A \code{crossomics_integration} object.}

\item{block}{Which block should be plotted.}

\item{axes}{Optional axes to include.}

\item{top_n}{Number of features to display.}
}
\value{
A \code{ggplot} object.
}
\description{
Display top loadings from one integration block as a heatmap-like tile plot.
}
