% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization_extended.R
\name{plot_differential_boxplot}
\alias{plot_differential_boxplot}
\title{Plot Differential Taxa as Boxplots}
\usage{
plot_differential_boxplot(
  object,
  microbiome_data,
  x = "sample_id",
  top_n = 12,
  relative = TRUE
)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object.}

\item{x}{Sample metadata column shown on the x axis.}

\item{top_n}{Number of taxa displayed.}

\item{relative}{Whether to use relative abundance.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw grouped boxplots for top differential taxa.
}
