% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization.R
\name{plot_differential_cladogram}
\alias{plot_differential_cladogram}
\title{Plot a Differential Abundance Cladogram}
\usage{
plot_differential_cladogram(object, top_n = 30)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}

\item{top_n}{Number of features retained for the cladogram.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a Differential Abundance Cladogram
}
