% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization.R
\name{plot_differential_effect_size}
\alias{plot_differential_effect_size}
\title{Plot Differential Effect Sizes}
\usage{
plot_differential_effect_size(object, top_n = 20)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}

\item{top_n}{Number of features displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Differential Effect Sizes
}
