% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization_extended.R
\name{plot_differential_ma}
\alias{plot_differential_ma}
\title{Plot a Differential Abundance MA Plot}
\usage{
plot_differential_ma(object, abundance_column = NULL, q_value_cutoff = 0.1)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}

\item{abundance_column}{Optional column in \code{object@result} used as abundance.}

\item{q_value_cutoff}{Significance cutoff.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw an MA-style plot of abundance versus effect size.
}
