% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization.R
\name{plot_differential_volcano}
\alias{plot_differential_volcano}
\title{Plot a Differential Abundance Volcano Plot}
\usage{
plot_differential_volcano(object, q_value_cutoff = 0.1, label_top_n = 10)
}
\arguments{
\item{object}{A \code{differential_abundance_result} object.}

\item{q_value_cutoff}{Significance threshold.}

\item{label_top_n}{Number of top features to label.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a Differential Abundance Volcano Plot
}
