% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_distance_heatmap}
\alias{plot_distance_heatmap}
\title{Plot a Distance Heatmap}
\usage{
plot_distance_heatmap(
  object,
  method = c("bray", "jaccard", "euclidean"),
  annotate_by = NULL,
  cluster = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} or beta-type \code{microbiome_diversity} object.}

\item{method}{Distance method used when \code{object} is a dataset.}

\item{annotate_by}{Optional sample metadata columns.}

\item{cluster}{Should samples be hierarchically clustered?}
}
\value{
A \code{ComplexHeatmap::Heatmap} object.
}
\description{
Visualize sample distances as a \code{ComplexHeatmap} heatmap.
}
\examples{
data("global_patterns", package = "microbiomedataset")
plot_distance_heatmap(global_patterns, annotate_by = "SampleType")
}
