% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a Microbiome Heatmap}
\usage{
plot_heatmap(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 20,
  relative = TRUE,
  cluster_samples = TRUE,
  cluster_taxa = TRUE,
  sample_label = "sample_id"
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{top_n}{Number of taxa displayed.}

\item{relative}{Whether to use relative abundance.}

\item{cluster_samples}{Whether to cluster samples.}

\item{cluster_taxa}{Whether to cluster taxa.}

\item{sample_label}{Sample metadata column shown on the x axis.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a heatmap of dominant taxa across samples.
}
\examples{
data("global_patterns", package = "microbiomedataset")

plot_heatmap(
  object = global_patterns,
  taxonomic_rank = "Phylum",
  top_n = 12,
  sample_label = "SampleType"
)
}
