% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization_extended.R
\name{plot_mechanism_summary}
\alias{plot_mechanism_summary}
\title{Plot a Mechanism Summary}
\usage{
plot_mechanism_summary(object, top_n = 20)
}
\arguments{
\item{object}{A \code{crossomics_mechanism} object.}

\item{top_n}{Number of links displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Rank taxon-pathway-metabolite mechanism links by score.
}
