% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization_extended.R
\name{plot_network_centrality}
\alias{plot_network_centrality}
\title{Plot Network Centrality}
\usage{
plot_network_centrality(
  object,
  metric = c("degree", "betweenness", "closeness"),
  top_n = 20
)
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}

\item{metric}{Centrality metric to display.}

\item{top_n}{Number of nodes displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize node centrality scores from a correlation network.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")
correlation <- calculate_correlation(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  microbiome_rank = "Genus",
  metabolome_transform = "none"
)
network <- build_correlation_network(
  correlation,
  min_abs_correlation = 0.2,
  max_q_value = 1,
  top_n = 20
)
plot_network_centrality(network)
}
