% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_result_visualization.R
\name{plot_network_modules}
\alias{plot_network_modules}
\title{Plot Network Modules}
\usage{
plot_network_modules(object, method = c("louvain", "walktrap"))
}
\arguments{
\item{object}{A \code{microbe_metabolite_network} object.}

\item{method}{Community detection method.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize network modules from a correlation network.
}
