% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot Taxon Prevalence}
\usage{
plot_prevalence(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  detection = 0,
  proportion = TRUE,
  top_n = 20
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{detection}{Detection threshold used in prevalence calculation.}

\item{proportion}{Whether prevalence should be shown as a proportion.}

\item{top_n}{Number of taxa displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a prevalence barplot for dominant taxa.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
plot_prevalence(
  object = x,
  taxonomic_rank = "Phylum",
  top_n = 10
)
}
