% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_prevalence_abundance}
\alias{plot_prevalence_abundance}
\title{Plot Prevalence versus Abundance}
\usage{
plot_prevalence_abundance(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  detection = 0,
  relative = TRUE,
  top_n = 20
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{detection}{Detection threshold used for prevalence.}

\item{relative}{Whether to use relative abundance.}

\item{top_n}{Number of taxa displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw a prevalence-abundance scatter plot for dominant taxa.
}
\examples{
data("global_patterns", package = "microbiomedataset")
x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
plot_prevalence_abundance(x, taxonomic_rank = "Phylum", top_n = 8)
}
