% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization.R
\name{plot_sample_depth}
\alias{plot_sample_depth}
\title{Plot Sample Depth}
\usage{
plot_sample_depth(object, x = "sample_id", geom = c("bar", "point"))
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{x}{Sample metadata column shown on the x axis.}

\item{geom}{Geometry used to display depth.}
}
\value{
A \code{ggplot} object.
}
\description{
Draw sequencing depth per sample.
}
\examples{
data("global_patterns", package = "microbiomedataset")

plot_sample_depth(global_patterns, x = "SampleType")
}
