% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_scree}
\alias{plot_scree}
\title{Plot an Ordination Scree Plot}
\usage{
plot_scree(
  object,
  method = c("PCA", "PCoA"),
  distance_method = c("bray", "jaccard", "euclidean"),
  top_n = 10
)
}
\arguments{
\item{object}{A \code{microbiome_ordination} or \code{microbiome_dataset} object.}

\item{method}{Ordination method used when \code{object} is a dataset.}

\item{distance_method}{Distance method used when computing PCoA.}

\item{top_n}{Number of axes displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Show explained variance across ordination axes.
}
\examples{
data("global_patterns", package = "microbiomedataset")
plot_scree(run_ordination(global_patterns, method = "PCA"))
}
