% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_taxa_boxplot}
\alias{plot_taxa_boxplot}
\title{Plot Taxa as Boxplots}
\usage{
plot_taxa_boxplot(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  taxa = NULL,
  top_n = 10,
  x = "sample_id",
  relative = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{taxa}{Optional taxa to keep. Defaults to the most abundant taxa.}

\item{top_n}{Number of taxa shown when \code{taxa} is not supplied.}

\item{x}{Sample metadata column shown on the x axis.}

\item{relative}{Whether to show relative abundance.}
}
\value{
A \code{ggplot} object.
}
\description{
Thin wrapper around \code{\link[=plot_abundance]{plot_abundance()}} with \code{geom = "boxplot"}.
}
\examples{
data("global_patterns", package = "microbiomedataset")
plot_taxa_boxplot(global_patterns, taxonomic_rank = "Phylum", x = "SampleType", top_n = 4)
}
