% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_taxonomy_sunburst}
\alias{plot_taxonomy_sunburst}
\title{Plot Taxonomy as a Sunburst}
\usage{
plot_taxonomy_sunburst(
  object,
  ranks = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus"),
  top_n = 50
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{ranks}{Taxonomy ranks used in the sunburst.}

\item{top_n}{Number of taxa displayed at the deepest rank.}
}
\value{
A \code{plotly} htmlwidget.
}
\description{
Draw a sunburst chart from hierarchical taxonomy abundances.
}
\examples{
data("global_patterns", package = "microbiomedataset")
plot_taxonomy_sunburst(global_patterns, ranks = c("Kingdom", "Phylum", "Class"))
}
