% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{plot_tree_heatmap}
\alias{plot_tree_heatmap}
\title{Plot a Tree-Ordered Heatmap}
\usage{
plot_tree_heatmap(
  object,
  tree = c("otu_tree", "taxa_tree"),
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 20,
  relative = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot used to derive row order.}

\item{taxonomic_rank}{Taxonomic rank used when \code{tree = "taxa_tree"}.}

\item{top_n}{Number of rows displayed when \code{tree = "taxa_tree"}.}

\item{relative}{Whether to use relative abundance.}
}
\value{
A \code{ComplexHeatmap::Heatmap} object.
}
\description{
Draw a heatmap ordered by the tip order of \code{taxa_tree} or \code{otu_tree}.
}
\examples{
data("global_patterns", package = "microbiomedataset")
x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- align_tree(x, tree = "taxa_tree")
plot_tree_heatmap(x, tree = "taxa_tree", taxonomic_rank = "Phylum")
}
