% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{plot_tree_link}
\alias{plot_tree_link}
\title{Plot Tree Link Coverage}
\usage{
plot_tree_link(
  object,
  tree = c("otu_tree", "taxa_tree"),
  x = c("link_rank", "tree")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{x}{Variable used on the x-axis. Defaults to \code{"link_rank"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize how many variables are linked to each tree rank or tree type.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
p <- plot_tree_link(x, tree = "taxa_tree")
class(p)
}
