% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{prepare_correlation_data}
\alias{prepare_correlation_data}
\title{Prepare Matched Data for Correlation Analysis}
\usage{
prepare_correlation_data(
  object = NULL,
  microbiome_data = NULL,
  metabolome_data = NULL,
  sample_link = NULL,
  microbiome_rank = "Genus",
  microbiome_transform = c("relative", "log", "presence_absence", "clr", "none"),
  metabolome_transform = c("log", "none"),
  min_taxa_prevalence = 0,
  min_taxa_abundance = 0,
  min_metabolite_prevalence = 0,
  min_metabolite_sd = 0
)
}
\arguments{
\item{object}{A \code{microbiome_metabolome_dataset} object.}

\item{microbiome_data}{A \code{microbiome_dataset} object. Used when \code{object} is
not supplied.}

\item{metabolome_data}{A \code{mass_dataset} object. Used when \code{object} is not
supplied.}

\item{sample_link}{Optional sample link table used when \code{object} is not
supplied.}

\item{microbiome_rank}{Optional taxonomy rank used to aggregate microbiome
abundance before extraction.}

\item{microbiome_transform}{Microbiome transformation method.}

\item{metabolome_transform}{Metabolome transformation method.}

\item{min_taxa_prevalence}{Minimum number of non-zero samples required for a
microbiome feature.}

\item{min_taxa_abundance}{Minimum mean abundance required for a microbiome
feature.}

\item{min_metabolite_prevalence}{Minimum number of non-zero samples required
for a metabolite.}

\item{min_metabolite_sd}{Minimum standard deviation required for a
metabolite.}
}
\value{
A list containing filtered matrices and metadata.
}
\description{
Build filtered microbiome and metabolome matrices suitable for correlation
analysis.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

x <- prepare_correlation_data(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  microbiome_rank = "Genus"
)
dim(x$microbiome_matrix)
}
