% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{prune_ref_seq}
\alias{prune_ref_seq}
\title{Prune Reference Sequences in a Microbiome Dataset}
\usage{
prune_ref_seq(object, variable_id, variable_index, keep = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{variable_id}{Variable identifiers to keep or remove.}

\item{variable_index}{Variable indices to keep or remove.}

\item{keep}{Should the provided sequences be kept? Defaults to \code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Keep or remove selected sequences from \code{ref_seq} without changing the
abundance matrix.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
x@ref_seq <- Biostrings::DNAStringSet(rep("ACGT", nrow(x@variable_info)))
names(x@ref_seq) <- x@variable_info$variable_id
x <- prune_ref_seq(x, variable_id = x@variable_info$variable_id[1:5])
length(x@ref_seq)
}
