% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{prune_samples}
\alias{prune_samples}
\title{Prune Samples from a Microbiome Dataset}
\usage{
prune_samples(object, sample_id, sample_index, keep = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{sample_id}{Sample identifiers to keep or remove.}

\item{sample_index}{Sample indices to keep or remove.}

\item{keep}{Should the provided samples be kept? Defaults to \code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Remove samples from a \code{microbiome_dataset} while keeping
\code{expression_data} and \code{sample_info} synchronized.
}
\examples{
data("global_patterns", package = "microbiomedataset")

keep_samples <- global_patterns@sample_info$sample_id[1:3]
x <- prune_samples(global_patterns, sample_id = keep_samples)
dim(x@expression_data)
}
