% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{prune_taxa}
\alias{prune_taxa}
\title{Prune Taxa from a Microbiome Dataset}
\usage{
prune_taxa(object, variable_id, variable_index, keep = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{variable_id}{Variable identifiers to keep or remove.}

\item{variable_index}{Variable indices to keep or remove.}

\item{keep}{Should the provided taxa be kept? Defaults to \code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Remove taxa from a \code{microbiome_dataset} while keeping
\code{expression_data}, \code{variable_info}, \code{otu_tree}, \code{taxa_tree}, and \code{ref_seq}
synchronized.
}
\examples{
data("global_patterns", package = "microbiomedataset")

keep_taxa <- global_patterns@variable_info$variable_id[1:5]
x <- prune_taxa(global_patterns, variable_id = keep_taxa)
dim(x@expression_data)
}
