% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{prune_tree}
\alias{prune_tree}
\title{Prune a Tree Slot in a Microbiome Dataset}
\usage{
prune_tree(object, tree = c("otu_tree", "taxa_tree"), tip_label, keep = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot to prune, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{tip_label}{Tip labels to keep or remove.}

\item{keep}{Should the provided tips be kept? Defaults to \code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Keep or remove selected tips from \code{otu_tree} or \code{taxa_tree} without changing
the abundance matrix.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:120])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x)
)
keep_tips <- x@variable_info$variable_id[1:5]
x <- prune_tree(x, tree = "otu_tree", tip_label = keep_tips)
class(x@otu_tree)
}
