% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abundance.R
\name{psmelt_microbiome_dataset}
\alias{psmelt_microbiome_dataset}
\title{Melt a microbiome_dataset into a Long Table}
\usage{
psmelt_microbiome_dataset(object, relative = FALSE, value_name = "abundance")
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{relative}{Should abundance be converted to relative abundance before
melting?}

\item{value_name}{Name of the abundance column in the output.}
}
\value{
A data frame in long format.
}
\description{
Convert a \code{microbiome_dataset} into a long table that combines abundance,
\code{sample_info}, and \code{variable_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- psmelt_microbiome_dataset(global_patterns)
head(x[, 1:8])

y <- psmelt_microbiome_dataset(global_patterns, relative = TRUE)
head(y[, c("sample_id", "variable_id", "abundance")])
}
