% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tree_seq.R
\name{replace_ref_seq}
\alias{replace_ref_seq}
\title{Replace Reference Sequences in a Microbiome Dataset}
\usage{
replace_ref_seq(object, value = NULL, align = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{value}{A \code{Biostrings::XStringSet} object or \code{NULL}.}

\item{align}{Should the sequences be aligned after replacement? Defaults to
\code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Set \code{ref_seq} on a \code{microbiome_dataset} and optionally align it with the
current dataset.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
ref_seq <- Biostrings::DNAStringSet(rep("ACGT", nrow(x@variable_info)))
x <- replace_ref_seq(x, value = ref_seq)
length(x@ref_seq)
}
