% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{replace_tree}
\alias{replace_tree}
\title{Replace a Tree Slot in a Microbiome Dataset}
\usage{
replace_tree(
  object,
  tree = c("otu_tree", "taxa_tree"),
  value = NULL,
  align = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{tree}{Tree slot to replace, either \code{"otu_tree"} or \code{"taxa_tree"}.}

\item{value}{A \code{treedata} object or \code{NULL}.}

\item{align}{Should the tree be aligned after replacement? Defaults to
\code{TRUE}.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Set \code{otu_tree} or \code{taxa_tree} on a \code{microbiome_dataset} and optionally align
it with the current dataset.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:80])
x <- microbiomedataset::convert2microbiome_dataset(
  microbiomedataset::convert2phyloseq(x0)
)
otu_phylo <- ape::rtree(
  n = nrow(x@variable_info),
  tip.label = x@variable_info$variable_id
)
x <- replace_tree(
  x,
  tree = "otu_tree",
  value = tidytree::treedata(
    phylo = otu_phylo,
    data = tibble::tibble(node = seq_len(length(otu_phylo$tip.label) + otu_phylo$Nnode))
  )
)
class(x@otu_tree)
}
