% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_microbiome.R
\name{run_ordination}
\alias{run_ordination}
\title{Run Ordination on a Microbiome Dataset}
\usage{
run_ordination(
  object,
  method = c("PCoA", "PCA"),
  distance_method = c("bray", "jaccard", "euclidean"),
  n_axes = 2
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{method}{Ordination method.}

\item{distance_method}{Distance method used for PCoA.}

\item{n_axes}{Number of axes to keep.}
}
\value{
A \code{microbiome_ordination} object.
}
\description{
Perform a basic ordination analysis and return an objectized result.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- run_ordination(global_patterns, method = "PCoA")
class(x)
head(x@sample_coord)
}
