% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{sample_sums}
\alias{sample_sums}
\title{Calculate Total Abundance per Sample}
\usage{
sample_sums(object, na.rm = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{na.rm}{Should missing values be removed before summation?}
}
\value{
A numeric vector named by \code{sample_id}.
}
\description{
Calculate Total Abundance per Sample
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- sample_sums(global_patterns)
head(x)
}
