% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{summarise_abundance}
\alias{summarise_abundance}
\title{Summarise Abundance by Taxa or Sample Groups}
\usage{
summarise_abundance(
  object,
  by = c("taxa", "samples"),
  group_by,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na_remove = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{by}{Summarisation target, either \code{"taxa"} or \code{"samples"}.}

\item{group_by}{Taxonomy rank or sample metadata column used for grouping.}

\item{what}{Aggregation method.}

\item{na_remove}{Should taxa with missing rank be removed before
summarisation when \code{by = "taxa"}?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Unified front-end for summarising abundance either across taxonomy ranks or
sample metadata groups.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
x <- summarise_abundance(
  object = x0,
  by = "taxa",
  group_by = "Genus",
  what = "sum_intensity"
)
dim(x@expression_data)

y <- summarise_abundance(
  object = x0,
  by = "samples",
  group_by = "SampleType",
  what = "sum_intensity"
)
dim(y@expression_data)
}
