% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{summarise_network_modules}
\alias{summarise_network_modules}
\title{Summarize Network Modules}
\usage{
summarise_network_modules(module_table)
}
\arguments{
\item{module_table}{Output from \code{\link[=detect_network_modules]{detect_network_modules()}}.}
}
\value{
A summary data.frame.
}
\description{
Aggregate module composition from a network module assignment table.
}
