% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_crossomics.R
\name{summarise_reactions}
\alias{summarise_reactions}
\title{Summarize Reaction-Level Mechanisms}
\usage{
summarise_reactions(mechanism_result, taxon_reaction_link = NULL)
}
\arguments{
\item{mechanism_result}{A \code{crossomics_mechanism} object.}

\item{taxon_reaction_link}{Optional taxon-to-reaction link table.}
}
\value{
A reaction summary data.frame.
}
\description{
Aggregate mechanism evidence from taxon-metabolite links to reaction-level
summaries, optionally incorporating taxon-to-reaction mappings.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")

pathway_link <- standardize_pathway_link(
  data.frame(
    taxon_id = summarise_taxa(
      demo_crossomics$microbiome_data,
      taxonomic_rank = "Genus"
    )@variable_info$variable_id[1],
    metabolite_id = demo_crossomics$metabolome_data@annotation_table$variable_id[1],
    pathway_id = "pathway_a",
    reaction_id = "reaction_a",
    reaction_name = "Reaction A",
    stringsAsFactors = FALSE
  )
)
mechanism <- infer_metabolic_link(
  microbiome_data = demo_crossomics$microbiome_data,
  metabolome_data = demo_crossomics$metabolome_data,
  sample_link = demo_crossomics$sample_link,
  pathway_link = pathway_link,
  microbiome_rank = "Genus",
  q_value_cutoff = 1
)
summarise_reactions(mechanism)
}
