% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_samples.R
\name{summarise_samples_by_group}
\alias{summarise_samples_by_group}
\title{Summarise Samples by a Sample Metadata Group}
\usage{
summarise_samples_by_group(
  object,
  group_by,
  what = c("sum_intensity", "mean_intensity", "median_intensity")
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{group_by}{A column in \code{sample_info}.}

\item{what}{Aggregation method.}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Microbiome-oriented wrapper around \code{\link[=summarise_samples]{summarise_samples()}} that summarizes
samples using a grouping column in \code{sample_info}.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- summarise_samples_by_group(
  object = global_patterns,
  group_by = "SampleType",
  what = "sum_intensity"
)
dim(x@expression_data)
}
