% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs_core.R
\name{summarise_taxa}
\alias{summarise_taxa}
\title{Summarise Taxa at a Taxonomic Rank}
\usage{
summarise_taxa(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na_remove = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{A taxonomy rank in \code{variable_info}.}

\item{what}{Aggregation method.}

\item{na_remove}{Should taxa with missing rank be removed before agglomeration?}
}
\value{
A \code{microbiome_dataset} object.
}
\description{
Microbiome-oriented wrapper around \code{\link[=agglomerate_taxa]{agglomerate_taxa()}} that summarizes
abundance at a chosen taxonomy rank.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
x <- summarise_taxa(
  object = x0,
  taxonomic_rank = "Family",
  what = "sum_intensity"
)
dim(x@expression_data)
}
