% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{taxa_prevalence}
\alias{taxa_prevalence}
\title{Calculate Taxon Prevalence across Samples}
\usage{
taxa_prevalence(object, detection = 0, proportion = FALSE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{detection}{Minimum abundance required to count as present.}

\item{proportion}{Return prevalence as a proportion instead of counts?}
}
\value{
A numeric vector named by \code{variable_id}.
}
\description{
Calculate Taxon Prevalence across Samples
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- taxa_prevalence(global_patterns, detection = 0)
head(x)
}
