% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_microbiome_dataset.R
\name{taxa_sums}
\alias{taxa_sums}
\title{Calculate Total Abundance per Taxon}
\usage{
taxa_sums(object, na.rm = TRUE)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{na.rm}{Should missing values be removed before summation?}
}
\value{
A numeric vector named by \code{variable_id}.
}
\description{
Calculate Total Abundance per Taxon
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- taxa_sums(global_patterns)
head(x)
}
