% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microbiome_visualization_extended.R
\name{time_series_taxa_plot}
\alias{time_series_taxa_plot}
\alias{plot_longitudinal}
\title{Plot Longitudinal Taxa Trajectories}
\usage{
plot_longitudinal(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  time_by,
  id_by = NULL,
  color_by = NULL,
  taxa = NULL,
  top_n = 6,
  relative = TRUE
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for aggregation.}

\item{time_by}{Sample metadata column used on the x axis.}

\item{id_by}{Optional metadata column used to connect repeated measures.}

\item{color_by}{Optional metadata column used for color.}

\item{taxa}{Optional taxa labels.}

\item{top_n}{Number of taxa shown when \code{taxa} is omitted.}

\item{relative}{Whether to use relative abundance.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot selected taxa across an ordered sample metadata variable.
}
\examples{
data("demo_crossomics", package = "microbiomedataset")
plot_longitudinal(
  demo_crossomics$microbiome_data,
  taxonomic_rank = "Genus",
  time_by = "age",
  id_by = "subject_id",
  top_n = 4
)
}
