% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_abundance.R
\name{transform2relative_intensity}
\alias{transform2relative_intensity}
\alias{transform2relative_intensity.microbiome_dataset}
\title{transform data to relative data}
\usage{
transform2relative_intensity(object, ...)

\method{transform2relative_intensity}{microbiome_dataset}(object, ...)
}
\arguments{
\item{object}{mass_dataset, microbiome_dataset or Taxa table}

\item{...}{other params}
}
\value{
mass_dataset, microbiome_dataset or Taxa table
}
\description{
transform data to relative data
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- transform2relative_intensity(global_patterns)
attr(x@expression_data, "counts_or_relative")
}
