% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_abundance.R
\name{transform_abundance_by}
\alias{transform_abundance_by}
\title{Transform Abundance by Taxa or Sample Groups}
\usage{
transform_abundance_by(
  object,
  by = c("taxa", "samples"),
  group_by,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  method = c("relative", "log", "presence_absence", "clr"),
  pseudocount = 1,
  na_remove = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{by}{Transformation target, either \code{"taxa"} or \code{"samples"}.}

\item{group_by}{Taxonomy rank or sample metadata column used for grouping.}

\item{what}{Aggregation method used before transformation.}

\item{method}{Transformation method.}

\item{pseudocount}{A pseudocount added before transformations that require
strictly positive values.}

\item{na_remove}{Should taxa with missing rank be removed before
summarisation when \code{by = "taxa"}?}

\item{...}{Other parameters passed to \code{\link[=transform_counts]{transform_counts()}}.}
}
\value{
A transformed \code{microbiome_dataset} object.
}
\description{
Unified front-end for transforming abundance after summarising either taxa or
samples.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
x <- transform_abundance_by(
  object = x0,
  by = "taxa",
  group_by = "Genus",
  what = "sum_intensity",
  method = "relative"
)
attr(x@expression_data, "transform_method")
}
