% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_abundance.R
\name{transform_counts}
\alias{transform_counts}
\alias{transform_counts.microbiome_dataset}
\title{Transform Count Data in a Microbiome Dataset}
\usage{
transform_counts(
  object,
  method = c("relative", "log", "presence_absence", "clr"),
  pseudocount = 1,
  ...
)

\method{transform_counts}{microbiome_dataset}(
  object,
  method = c("relative", "log", "presence_absence", "clr"),
  pseudocount = 1,
  ...
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{method}{Transformation method.}

\item{pseudocount}{A pseudocount added before transformations that require
strictly positive values.}

\item{...}{Other parameters.}
}
\value{
A transformed \code{microbiome_dataset} object.
}
\description{
Transform Count Data in a Microbiome Dataset
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- transform_counts(global_patterns, method = "relative")
attr(x@expression_data, "counts_or_relative")

y <- transform_counts(global_patterns, method = "log", pseudocount = 1)
attr(y@expression_data, "transform_method")
}
