% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_abundance.R
\name{transform_taxa}
\alias{transform_taxa}
\title{Transform Taxa after Taxonomic Summarization}
\usage{
transform_taxa(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  method = c("relative", "log", "presence_absence", "clr"),
  pseudocount = 1,
  na_remove = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object.}

\item{taxonomic_rank}{Taxonomic rank used for summarization.}

\item{what}{Aggregation method used before transformation.}

\item{method}{Transformation method.}

\item{pseudocount}{A pseudocount added before transformations that require
strictly positive values.}

\item{na_remove}{Should taxa with missing rank be removed before
summarization?}

\item{...}{Other parameters passed to \code{\link[=transform_counts]{transform_counts()}}.}
}
\value{
A transformed \code{microbiome_dataset} object.
}
\description{
Summarize a \code{microbiome_dataset} at a chosen taxonomy rank and then apply a
count transformation to the aggregated abundance matrix.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x0 <- prune_taxa(global_patterns, variable_id = global_patterns@variable_info$variable_id[1:200])
x <- transform_taxa(
  object = x0,
  taxonomic_rank = "Genus",
  what = "sum_intensity",
  method = "relative"
)
attr(x@expression_data, "transform_method")
}
