% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_microbiome_dataset.R
\name{update_microbiome_dataset}
\alias{update_microbiome_dataset}
\title{Upgrade a microbiome_dataset Object to the Current Schema}
\usage{
update_microbiome_dataset(object)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object, including objects saved with an
older class definition.}
}
\value{
A refreshed \code{microbiome_dataset} object.
}
\description{
Rebuild a \code{microbiome_dataset} using the current class definition and derive
missing explicit tree link metadata.
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <- update_microbiome_dataset(global_patterns)
x@version
}
