% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_microbiome_dataset.R
\docType{methods}
\name{activate_microbiome_dataset}
\alias{activate_microbiome_dataset}
\title{activate_microbiome_dataset}
\usage{
activate_microbiome_dataset(
  .data,
  what = c("sample_info", "variable_info", "expression_data")
)
}
\arguments{
\item{.data}{(required) A microbiome_dataset class object}

\item{what}{(required) What should get be activated?
Possible values are \code{sample_info},
\code{expression_data}, or \code{variable_info}.}
}
\value{
A microbiome_dataset class object
}
\description{
Determine the context of subsequent manipulations
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
