% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_microbiome_dataset.R
\docType{methods}
\name{check_microbiome_dataset}
\alias{check_microbiome_dataset}
\title{check_microbiome_dataset}
\usage{
check_microbiome_dataset(
  expression_data,
  sample_info,
  variable_info,
  otu_tree,
  taxa_tree,
  ref_seq,
  sample_info_note,
  variable_info_note
)
}
\arguments{
\item{expression_data}{expression_data.}

\item{sample_info}{sample_info.}

\item{variable_info}{variable_info.}

\item{otu_tree}{otu_tree}

\item{taxa_tree}{taxa_tree}

\item{ref_seq}{ref_seq}

\item{sample_info_note}{sample_info_note.}

\item{variable_info_note}{variable_info_note.}
}
\value{
Notice of data checking.
}
\description{
Check data format for microbiome_dataset class object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
