% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_microbiome_dataset_class.R
\docType{methods}
\name{check_microbiome_dataset_class}
\alias{check_microbiome_dataset_class}
\title{check_microbiome_dataset_class}
\usage{
check_microbiome_dataset_class(object)
}
\arguments{
\item{object}{microbiome_dataset class.}
}
\value{
Notice of data checking.
}
\description{
Check microbiome_dataset class object.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
