% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_convert2mass_dataset.R
\name{convert2mass_dataset}
\alias{convert2mass_dataset}
\alias{as.mass_dataset}
\alias{convert2mass_dataset.microbiome_dataset}
\title{convert to mass_dataset class object.}
\usage{
convert2mass_dataset(object, ...)

as.mass_dataset(object, ...)

\method{convert2mass_dataset}{microbiome_dataset}(object, ...)
}
\arguments{
\item{object}{microbiomedataset}

\item{...}{other params}
}
\value{
mass_dataset class object.
}
\description{
convert to mass_dataset class object.
}
