% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_convert2microbiome_dataset.R
\name{convert2microbiome_dataset}
\alias{convert2microbiome_dataset}
\alias{as.microbiome_dataset}
\alias{convert2microbiome_dataset.phyloseq}
\alias{convert2microbiome_dataset.mass_dataset}
\title{convert to microbiome_dataset class object.}
\usage{
convert2microbiome_dataset(object, ...)

as.microbiome_dataset(object, ...)

\method{convert2microbiome_dataset}{phyloseq}(object, ...)

\method{convert2microbiome_dataset}{mass_dataset}(object, ...)
}
\arguments{
\item{object}{phylosep or other class}

\item{...}{other params}
}
\value{
microbiome_dataset class object.
}
\description{
convert to microbiome_dataset class object.
}
