% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_convert2treedata.R
\name{convert2treedata}
\alias{convert2treedata}
\alias{as.treedata}
\alias{convert2treedata.data.frame}
\alias{convert2treedata.taxonomyTable}
\title{convert data.frame contained hierarchical relationship or other classes to treedata class}
\usage{
convert2treedata(object, type = "species", ...)

as.treedata(object, type = "species", ...)

\method{convert2treedata}{data.frame}(object, type = "species", ...)

\method{convert2treedata}{taxonomyTable}(object, type = "species", ...)
}
\arguments{
\item{object}{taxonomyTable class (phyloseq package) or other class (data.frame)}

\item{type}{type}

\item{...}{other params}
}
\value{
treedata class object (ggtree package).
}
\description{
convert data.frame contained hierarchical relationship or other classes to treedata class
}
