% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extract.R
\docType{methods}
\name{extract_otu_tree}
\alias{extract_otu_tree}
\alias{extract_taxa_tree}
\alias{extract_ref_seq}
\title{extract_otu_tree}
\usage{
extract_otu_tree(object)

extract_taxa_tree(object)

extract_ref_seq(object)
}
\arguments{
\item{object}{(required) microbiome_dataset class object.}
}
\value{
A treedata

A treedata

A xstringset
}
\description{
Extract OTU tree.

Extract Taxa tree.

Extract ref sequence.
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
