% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_extract_intensity.R
\name{extract_intensity}
\alias{extract_intensity}
\alias{extract_intensity.microbiome_dataset}
\title{extract_intensity}
\usage{
extract_intensity(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus",
    "Species"),
  data_type = c("wider", "longer"),
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  ...
)

\method{extract_intensity}{microbiome_dataset}(
  object,
  taxonomic_rank = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus",
    "Species"),
  data_type = c("longer", "wider"),
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{taxonomic_rank}{taxonomic_rank}

\item{data_type}{if return_same_object is FALSE, the return data.frame}

\item{what}{which you want to mutate}

\item{na.rm}{na.rm}

\item{relative}{relative intensity or not}

\item{...}{other params}
}
\value{
microbiome_dataset
}
\description{
extract_intensity
}
\examples{
data("global_patterns", package = "microbiomedataset")

x <-
  extract_intensity(
    object = global_patterns,
    taxonomic_rank = "Genus",
    data_type = "longer",
    relative = TRUE
  )
head(x)
x <-
  extract_intensity(
    object = global_patterns,
    taxonomic_rank = "Genus",
    data_type = "wider",
    relative = FALSE,
    what = "sum_intensity"
  )
head(x[, 1:5])
}
