% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_microbiomedataset_logo.R
\docType{methods}
\name{get_microbiomedataset_version}
\alias{get_microbiomedataset_version}
\title{get_microbiomedataset_version}
\usage{
get_microbiomedataset_version()
}
\value{
version
}
\description{
Get microbiomedataset package version
}
\examples{
get_microbiomedataset_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
