% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_in_package.R
\docType{data}
\name{global_patterns}
\alias{global_patterns}
\alias{GlobalPatterns}
\title{(Data) Global patterns of 16S rRNA diversity at a depth of millions of sequences per sample (2011)}
\description{
Published in PNAS in early 2011. This work compared the microbial
communities from 25 environmental samples and three known “mock communities”
– a total of 9 sample types – at a depth averaging 3.1 million reads per
sample. Authors were able to reproduce diversity patterns seen in many
other published studies, while also invesitigating technical
issues/bias by applying the same techniques to simulated microbial
communities of known composition.
This data is from phyloseq package
}
\examples{
data(global_patterns)
}
\keyword{data}
