% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{microbiome_dataset-class}
\alias{microbiome_dataset-class}
\title{microbiome_dataset class}
\description{
microbiome_dataset class
}
\section{Slots}{

\describe{
\item{\code{otu_tree}}{A treedata object of tidytree package or NULL.}

\item{\code{taxa_tree}}{A treedata object of tidytree package or NULL.}

\item{\code{ref_seq}}{A XStringSet object of Biostrings package or NULL.}

\item{\code{...}}{Other slots from \code{\link[massdataset:mass_dataset]{mass_dataset}}}
}}

