% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_microbiomedataset_logo.R
\docType{methods}
\name{microbiomedataset_logo}
\alias{microbiomedataset_logo}
\title{microbiomedataset_logo}
\usage{
microbiomedataset_logo()
}
\value{
logo
}
\description{
Get the detailed information of microbiomedataset package.
}
\examples{
microbiomedataset_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
