% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mutate.R
\name{mutate2sample}
\alias{mutate2sample}
\alias{mutate2sample.microbiome_dataset}
\title{Mutate new columns in sample_info}
\usage{
mutate2sample(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number",
    "na_prop"),
  group_by,
  na.rm = TRUE,
  relative = TRUE,
  data_type = c("longer", "wider"),
  return_same_object = TRUE,
  ...
)

\method{mutate2sample}{microbiome_dataset}(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number",
    "na_prop"),
  group_by,
  na.rm = TRUE,
  relative = TRUE,
  data_type = c("longer", "wider"),
  return_same_object = TRUE,
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{what}{which you want to mutate}

\item{group_by}{should be one column from variable_info}

\item{na.rm}{na.rm}

\item{relative}{relative intensity or not.}

\item{data_type}{if return_same_object is FALSE, the return data.frame is
longer or wider.}

\item{return_same_object}{return same object or not.}

\item{...}{other params}
}
\value{
microbiome_dataset or data.frame.
}
\description{
Mutate new columns in sample_info
}
