% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_mutate2variable.R
\name{mutate2variable}
\alias{mutate2variable}
\alias{mutate2variable.microbiome_dataset}
\title{Mutate New Columns into Variable Info of Microbiome Dataset}
\usage{
mutate2variable(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop",
    "zero_number", "zero_prop"),
  according_to_samples = "all",
  ...
)

\method{mutate2variable}{microbiome_dataset}(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity", "na_number", "na_prop",
    "zero_number", "zero_prop"),
  according_to_samples = "all",
  ...
)
}
\arguments{
\item{object}{A \code{microbiome_dataset} object on which the mutation will be performed.}

\item{what}{A character vector specifying the statistics to be calculated and added.
Options include "mean_intensity", "median_intensity", "sum_intensity", "na_number",
"na_prop", "zero_number", "zero_prop". Default is all options.}

\item{according_to_samples}{A character vector specifying the samples to be used for
calculations. Default is "all", which uses all samples. If only specific samples are
needed, provide their names as a vector.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
Returns the modified \code{microbiome_dataset} object with additional columns in
\code{variable_info}.
}
\description{
This function adds new columns to the \code{variable_info} of a \code{microbiome_dataset}.
It calculates various statistics like mean, median, sum of intensities, and proportions
of NA and zero values, based on the provided samples.
}
