% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dplyr.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{select}
\alias{desc}
\alias{arrange}
\alias{count}
\alias{glimpse}
\alias{group_by}
\alias{left_join}
\alias{mutate}
\alias{transmute}
\alias{mutate_all}
\alias{mutate_at}
\alias{n}
\alias{pull}
\alias{relocate}
\alias{rename}
\alias{slice}
\alias{slice_head}
\alias{slice_tail}
\alias{slice_sample}
\alias{slice_min}
\alias{slice_max}
\alias{summarize}
\alias{check_mass_dataset_class}
\alias{check_mass_dataset}
\alias{ungroup}
\alias{distinct}
\alias{nest}
\alias{unnest}
\alias{\%>\%}
\alias{extract_expression_data}
\alias{extract_sample_info}
\alias{extract_variable_info}
\alias{extract_variable_info_note}
\alias{extract_sample_info_note}
\alias{extract_process_info}
\alias{report_parameters}
\alias{parse_tidymass_parameter}
\alias{get_mv_number}
\alias{get_sample_id}
\alias{get_sample_number}
\alias{get_variable_id}
\alias{get_variable_number}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{glimpse}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{massdataset}{\code{\link[massdataset]{check_mass_dataset}}, \code{\link[massdataset]{check_mass_dataset_class}}, \code{\link[massdataset:extract-mass_dataset]{extract_expression_data}}, \code{\link[massdataset:extract-mass_dataset]{extract_process_info}}, \code{\link[massdataset:extract-mass_dataset]{extract_process_info}}, \code{\link[massdataset:extract-mass_dataset]{extract_sample_info}}, \code{\link[massdataset:extract-mass_dataset]{extract_sample_info_note}}, \code{\link[massdataset:extract-mass_dataset]{extract_variable_info}}, \code{\link[massdataset:extract-mass_dataset]{extract_variable_info_note}}, \code{\link[massdataset:summary-mass_dataset]{get_mv_number}}, \code{\link[massdataset:summary-mass_dataset]{get_sample_id}}, \code{\link[massdataset:summary-mass_dataset]{get_sample_number}}, \code{\link[massdataset:summary-mass_dataset]{get_variable_id}}, \code{\link[massdataset:summary-mass_dataset]{get_variable_number}}, \code{\link[massdataset]{parse_tidymass_parameter}}, \code{\link[massdataset]{report_parameters}}}

  \item{tidyr}{\code{\link[tidyr]{nest}}, \code{\link[tidyr]{unnest}}}
}}

