% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summarise_variables.R
\name{summarise_variables}
\alias{summarise_variables}
\alias{summarize_variables}
\alias{summarise_variables.microbiome_dataset}
\title{merge samples}
\usage{
summarise_variables(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  group_by,
  ...
)

summarize_variables(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  group_by,
  ...
)

\method{summarise_variables}{microbiome_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  group_by,
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{what}{which you want to use}

\item{variable_id}{variables should be merged}

\item{variable_index}{variables should be merged}

\item{remain_variable_info_id}{which variable information should be used.}

\item{remain_variable_info_index}{which variable information should be used.}

\item{group_by}{group all variables by column in variable_info}

\item{...}{other params}
}
\value{
microbiome_dataset
}
\description{
merge samples
}
\examples{
data(global_patterns)
dim(global_patterns)
global_patterns

global_patterns <-
  global_patterns \%>\%
  activate_microbiome_dataset(what = "variable_info") \%>\%
  dplyr::filter(!is.na(Genus))

object <-
  global_patterns \%>\%
  summarize_variables(what = "sum_intensity", group_by = "Genus")

dim(object)
}
