% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summarise_variables.R
\name{summarise_variables}
\alias{summarise_variables}
\alias{summarize_variables}
\alias{summarise_variables.microbiome_dataset}
\title{merge samples}
\usage{
summarise_variables(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  ...
)

summarize_variables(
  object,
  what = c("mean_intensity", "median_intensity", "sum_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  ...
)

\method{summarise_variables}{microbiome_dataset}(
  object,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  variable_id,
  variable_index,
  remain_variable_info_id,
  remain_variable_info_index,
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{what}{which you want to use}

\item{variable_id}{variables should be merged}

\item{variable_index}{variables should be merged}

\item{remain_variable_info_id}{which variable information should be used.}

\item{remain_variable_info_index}{which variable information should be used.}

\item{...}{other params}
}
\value{
microbiome_dataset
}
\description{
merge samples
}
