% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_transform.R
\name{transform2relative_intensity}
\alias{transform2relative_intensity}
\alias{transform2relative_intensity.microbiome_dataset}
\title{transform data to relative data}
\usage{
transform2relative_intensity(object, ...)

\method{transform2relative_intensity}{microbiome_dataset}(object, ...)
}
\arguments{
\item{object}{mass_dataset, microbiome_dataset or Taxa table}

\item{...}{other params}
}
\value{
mass_dataset, microbiome_dataset or Taxa table
}
\description{
transform data to relative data
}
